+++ event-notificator/src/main/java/ru/haritonenko/eventnotificator/api/dto/EventNotificationDto.java
@@ -3,8 +3,8 @@ package ru.haritonenko.eventnotificator.api.dto;
 import java.time.LocalDateTime;
 
 public record EventNotificationDto(
-        Integer id,
-        Integer eventId,
+        Long id,
+        Long eventId,
         LocalDateTime createdAt,
         boolean read,
         String message
